backend/requirements.txt
Flask
Flask-Cors
python-dotenv
requests
pandas
yfinance
ollama
pillow
google-genai
lxml
stock-indicators
openai